requirements.txt
annotated-types
anyio
attrs
beautifulsoup4
cachetools
certifi
cffi
charset-normalizer
click
curl_cffi
dateutils
distro
et_xmlfile
filetype
frozendict
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httpx
httpx-sse
idna
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-core
langchain-google-genai
langchain-mcp-adapters
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
mcp
multitasking
nsepython
nsetools
numpy
openai
openpyxl
orjson
ormsgpack
packaging
pandas
peewee
platformdirs
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
regex
requests
requests-toolbelt
rpds-py
rsa
scipy
setuptools
six
sniffio
soupsieve
SQLAlchemy
sse-starlette
starlette
tenacity
tiktoken
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
websockets
xlrd
xxhash
yfinance
zstandard
langchain-community
redis
fastapi[standard]